[IonutCava] - Fix a few bugs related to geometry loading and glVertexArray offsets -- Also, optimise DVDConverter a bit - Add a conditional toggle for texture state flushing (thus allowing us to avoid generating mipmaps for shadowmaps before the blur step and again after it) - Add an option in config.h to disable/enable multi-context resource loading - Added SFINAE type checks to ResourceCache class
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@515 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 09457d9e95f2381491b2556b3558fd235194c5a2 [formerly 5fcc006242e855fbecc6eee68ea813eccaa6469a] [formerly c0af4d69b966938a98cae2d00d68b105397b3f3e [formerly 22408864527ed1fcc0226aa2a3c5b8ea62caa700]]
Former-commit-id: b3bb719ce0d9031c802e286b2ac1d9dfd8d695e7 [formerly 92f0cdb676c54bc5c2566cb72c1c880808e888a2]
Former-commit-id: 26e73dc02abaeee674b1ba25f5337e25afee5795

+++ Source Code/Platform/Video/OpenGL/Buffers/VertexBuffer/Headers/glVertexArray.h
@@ -104,7 +104,7 @@ class glVertexArray : public VertexBuffer {
     ByteBuffer _smallData;
     typedef std::array<bool, to_const_uint(VertexAttribute::COUNT)> AttribFlags;
     AttribFlags _useAttribute;
-    typedef std::array<size_t, to_const_uint(VertexAttribute::COUNT)> AttribValues;
+    typedef std::array<GLuint, to_const_uint(VertexAttribute::COUNT)> AttribValues;
     AttribValues _attributeOffset;
 
     size_t _vaoHash;
